+++ crates/memory/src/arc/ptr_mut.rs
@@ -1,7 +1,10 @@
 use std::{
-    cell::{Ref, RefCell, RefMut},
     fmt,
     ops::{Deref, DerefMut},
+};
+
+use parking_lot::{
+    MappedRwLockReadGuard, MappedRwLockWriteGuard, RwLock, RwLockReadGuard, RwLockWriteGuard,
 };
 
 use crate::Ptr;
@@ -15,9 +18,9 @@ use crate::Ptr;
 #[macro_export]
 macro_rules! make_ptr_mut {
     ($value:expr) => {{
-        use std::rc::Rc;
+        use std::sync::Arc;
 
-        PtrMut::from(Rc::from(KCell::from($value)) as Rc<KCell<_>>)
+        PtrMut::from(Arc::from(KCell::from($value)) as Arc<KCell<_>>)
     }};
 }
 
@@ -32,11 +35,11 @@ impl<T> From<T> for PtrMut<T> {
 
 /// A mutable value with borrowing checked at runtime
 #[derive(Debug, Default)]
-pub struct KCell<T: ?Sized>(RefCell<T>);
+pub struct KCell<T: ?Sized>(RwLock<T>);
 
 impl<T> From<T> for KCell<T> {
     fn from(value: T) -> Self {
-        Self(RefCell::from(value))
+        Self(RwLock::from(value))
     }
 }
 
@@ -48,14 +51,14 @@ impl<T: ?Sized> KCell<T> {
     /// If the value is currently mutably borrowed then this function will block.
     /// See `try_borrow` for a non-blocking version.
     pub fn borrow(&self) -> Borrow<T> {
-        Borrow::new(self.0.borrow())
+        Borrow::new(self.0.read())
     }
 
     /// Attempts to mutably borrow the wrapped value.
     ///
     /// Returns an error if the value is currently mutably borrowed.
     pub fn try_borrow(&self) -> Option<Borrow<'_, T>> {
-        self.0.try_borrow().ok().map(Borrow::new)
+        self.0.try_read().map(Borrow::new)
     }
 
     /// Mutably borrows the wrapped value.
@@ -64,23 +67,23 @@ impl<T: ?Sized> KCell<T> {
     /// locked.
     /// See `try_borrow_mut` for a non-blocking version.
     pub fn borrow_mut(&self) -> BorrowMut<T> {
-        BorrowMut::new(self.0.borrow_mut())
+        BorrowMut::new(self.0.write())
     }
 
     /// Attempts to mutably borrow the wrapped value.
     ///
     /// Returns an error if the value is currently mutably borrowed.
     pub fn try_borrow_mut(&self) -> Option<BorrowMut<'_, T>> {
-        self.0.try_borrow_mut().ok().map(BorrowMut::new)
+        self.0.try_write().map(BorrowMut::new)
     }
 }
 
-/// An immutably borrowed reference to a value borrowed from a [PtrMut]
-pub struct Borrow<'a, T: ?Sized>(Ref<'a, T>);
+/// An immutably borrowed reference to a value borrowed from a [KCell]
+pub struct Borrow<'a, T: ?Sized>(MappedRwLockReadGuard<'a, T>);
 
 impl<'a, T: ?Sized> Borrow<'a, T> {
-    fn new(guard: Ref<'a, T>) -> Self {
-        Self(Ref::map(guard, |x| x))
+    fn new(guard: RwLockReadGuard<'a, T>) -> Self {
+        Self(RwLockReadGuard::map(guard, |x| x))
     }
 
     /// Makes a new Borrow for an optional component of the borrowed data.
@@ -90,7 +93,9 @@ impl<'a, T: ?Sized> Borrow<'a, T> {
         F: FnOnce(&T) -> Option<&U>,
         U: ?Sized,
     {
-        Ref::filter_map(borrowed.0, f).map(Borrow).map_err(Borrow)
+        MappedRwLockReadGuard::try_map(borrowed.0, f)
+            .map(Borrow)
+            .map_err(Borrow)
     }
 }
 
@@ -109,12 +114,12 @@ impl<T: ?Sized + fmt::Display> fmt::Display for Borrow<'_, T> {
     }
 }
 
-/// A mutably borrowed reference to a value borrowed from a [PtrMut]
-pub struct BorrowMut<'a, T: ?Sized>(RefMut<'a, T>);
+/// A mutably borrowed reference to a value borrowed from a [KCell]
+pub struct BorrowMut<'a, T: ?Sized>(MappedRwLockWriteGuard<'a, T>);
 
 impl<'a, T: ?Sized> BorrowMut<'a, T> {
-    fn new(guard: RefMut<'a, T>) -> Self {
-        Self(RefMut::map(guard, |x| x))
+    fn new(guard: RwLockWriteGuard<'a, T>) -> Self {
+        Self(RwLockWriteGuard::map(guard, |x| x))
     }
 
     /// Makes a new BorrowMut for an optional component of the borrowed data.
@@ -124,7 +129,7 @@ impl<'a, T: ?Sized> BorrowMut<'a, T> {
         F: FnOnce(&mut T) -> Option<&mut U>,
         U: ?Sized,
     {
-        RefMut::filter_map(borrowed.0, f)
+        MappedRwLockWriteGuard::try_map(borrowed.0, f)
             .map(BorrowMut)
             .map_err(BorrowMut)
     }