+++ crates/runtime/src/io/file.rs
@@ -1,7 +1,7 @@
-use crate::{runtime_error, KString, Result};
+use crate::{prelude::*, Result};
 
 /// A trait used for file-like-things in Koto
-pub trait KotoFile: KotoRead + KotoWrite {
+pub trait KotoFile: KotoRead + KotoWrite + KotoSend + KotoSync {
     /// An identifier for the file, accessed when displaying the file in strings
     fn id(&self) -> KString;